Moved Building_DroneCultivator.cs
